apply api search and link from suggestion and search page to page
apply api search and link from suggestion and search page to page
